==> CITE-seq/08-sub_clustering/CD4_T/01-CD4_T_cells_subset_by_cluster.Rmd <==
---
title: "CD4+ T cell clusters"
date: "`r format(Sys.time(), '%d %B, %Y')`"
author: Sonal 
output:
  html_document:
    toc: yes
    toc_float: yes
    number_sections: yes
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(knitr)
setwd("/home/srashmi/Documents/tonsil_atlas_citeseq_vdj_20210505/")
```

# Define parameters

## Load packages

```{r } 
library(Seurat)
library(flexclust)
library(tidyverse)
library(dplyr)
library(ggplot2)
library(patchwork)
library(scales)
library(harmony)
```


## Define variables and functions

```{r}
path_to_save_obj <- "/home/srashmi/Documents/tonsil_atlas_citeseq_vdj_20210505/results"
path_to_save_citeseq_seurat_obj <- str_c(
  path_to_save_obj,
  "tonsil_cite_seq_annotated_filtered.rds",
  sep = "/"
)
path_to_save_seurat_obj <- str_c(
  "objects/seurat_object_cite_seq_cd4_positive_object.rds",
  sep = "/"
)
```

## Load data

```{r data}
seurat_obj <- readRDS(path_to_save_citeseq_seurat_obj)
```

## get metadata

```{r get metadata}
metadata <- seurat_obj@meta.data 
```

# extract CD4+ cells

```{r}
CD4pos_clusters_obj = subset(seurat_obj,idents = c("2"))
```


# QC of CD4+ enriched clusters

```{r}
DimPlot(CD4pos_clusters_obj)
```

```{r}
FeaturePlot(CD4pos_clusters_obj, features = c("CD8","CD3","CD4.1"), ncol = 2)
```


# Save the seurat object 

```{r}
saveRDS(CD4pos_clusters_obj, file = path_to_save_seurat_obj)
```

# Session Information

```{r session}
sessionInfo()
```
