==> CITE-seq/08-sub_clustering/CD4_T/02-project_citeseq_onto_rna.Rmd <==
---
title: "Project CD4 CITE-seq onto RNA-seq (level 5)"
date: "`r format(Sys.time(), '%B %Y')`"
output:
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center',
                      message=FALSE, warning = FALSE)
options(width = 1200)
```


# Introduction

This script predicts transfers the annotation (level 5) and UMAP coordinates from scRNA-seq to CITE-seq.


# Pre-processing

## Load packages

```{r}
library(Seurat)
#library(SeuratWrappers)
library(harmony)
library(caret)
library(class)
library(tidyverse)
```


## Parameters

```{r}
# Parameters and thresholds
optimal_k <- 8
# Source functions
source("SLOcatoR_functions.R")
# Paths
path_to_query <- "objects/seurat_object_cite_seq_cd4_positive_object.rds"
path_to_reference <- "objects/CD4_T_subseted_annotated_level_5.rds"
path_to_save_cite_seq_df <- "objects/umap_df_cite_seq.tsv"
path_to_save_integrated <- "objects/CD4_T_cell_integrated_cite_seq_rna.rds"
```


## Read data

```{r}
query <- readRDS(path_to_query)
reference <- readRDS(path_to_reference)
```


# Merge query and reference

```{r}
DefaultAssay(query) <- "RNA"
query[["ADT"]] <- NULL
reference@meta.data$subproject <- reference@meta.data$donor_id 
reference@meta.data$UMAP_1_level_5 <- reference@reductions$umap@cell.embeddings[,1]
reference@meta.data$UMAP_2_level_5 <- reference@reductions$umap@cell.embeddings[,2]
reference@meta.data$annotation_level_5 <- reference@meta.data$annotation_paper
reference <- DietSeurat(reference)
selected_cols_ref <- c("donor_id","gem_id", "assay", "nCount_RNA","nFeature_RNA", "pct_mt","annotation_level_5",     "scrublet_doublet_scores", "scrublet_predicted_doublet", "UMAP_1_level_5", "UMAP_2_level_5")
reference@meta.data <- reference@meta.data[, selected_cols_ref]
reference$type <- "reference"
colnames(query@meta.data)[colnames(query@meta.data) == "gemid"] <- "gem_id" 
colnames(query@meta.data)[colnames(query@meta.data) == "percent.mt"] <- "pct_mt" 
selected_cols_query <- c("donor_id","gem_id", "nCount_RNA","nFeature_RNA","pct_mt","scrublet_doublet_scores", "scrublet_predicted_doublet")
query@meta.data <- query@meta.data[, selected_cols_query]
query$assay <- "CITE-seq"
query$type <- "query"
query@meta.data$annotation_level_5 <- "unannotated"
query$UMAP_1_level_5 <- NA
query$UMAP_2_level_5 <- NA
query@meta.data$UMAP_1_level_5 <- query$UMAP_1_level_5
query@meta.data$UMAP_2_level_5 <- query$UMAP_2_level_5
integrated <- merge(x = reference, y = query)
Idents(integrated) <- "assay"
```


# Find assay-specific highly variable genes

```{r}
shared_hvg <- find_assay_specific_features(integrated, n_features = 5500)
print(length(shared_hvg))
```


# Integrate with Harmony

```{r}
integrated <- integrate_assays(
  integrated,
  assay_specific = TRUE,
  assay_var = "assay",
  shared_hvg = shared_hvg
)
integrated <- RunUMAP(integrated, dims = 1:30, reduction = "harmony")
DimPlot(integrated, split.by = "assay", group.by = "assay")
VlnPlot(integrated, "nFeature_RNA", group.by = "assay", pt.size = 0)
```


# Create training and test sets

```{r}
data_sets <- split_training_and_test_sets(
  integrated,
  split_var = "type",
  referece_label = "reference",
  query_label = "query",
  reduction = "harmony",
  n_dims = 30
)
```


# Find optimal K

```{r}
optimal_k_df <- find_optimal_k(
  seurat_obj = integrated,
  training_set = data_sets$training_set,
  response_var = "annotation_level_5",
  return_plot = TRUE
)
optimal_k_df$df
optimal_k_df$plot
```


# Transfer label

```{r}
annotation_data  <- transfer_label(
  seurat_obj = integrated,
  training_set = data_sets$training_set,
  test_set = data_sets$test_set,
  k = optimal_k,
  response_var = "annotation_level_5"
)
```


# Predict UMAP coords

```{r}
umap_coords <- transfer_umap_coords(
  seurat_obj = integrated,
  training_set = data_sets$training_set,
  test_set = data_sets$test_set,
  k = optimal_k,
  umap1_var = "UMAP_1_level_5",
  umap2_var = "UMAP_2_level_5"
)
ggplot(umap_coords, aes(UMAP1, UMAP2)) +
  geom_point() +
  theme_classic()
```


# Save

```{r}
# Homogenize columns

integrated@meta.data[annotation_data$query_cell,]$annotation_level_5 <- "NA"
integrated@meta.data[annotation_data$query_cell,]$annotation_level_5 <- annotation_data$annotation_level_5
#integrated@meta.data[annotation_data$query_cell,]$annotation_prob <- "NA"
#integrated@meta.data[annotation_data$query_cell,]$annotation_prob <- annotation_data$annotation_prob
#integrated$annotation_level_1 <- factor(integrated$annotation_level_1)
#list <- unique(integrated$annotation_level_1)
#if (1 %in% list){
#  integrated@meta.data[annotation_data$query_cell,]$annotation_level_1 <- "NA"
#  integrated@meta.data[annotation_data$query_cell,]$annotation_level_1 <- annotation_data$annotation_level_1
#  integrated@meta.data[annotation_data$query_cell,]$annotation_prob <- "NA"
#  integrated@meta.data[annotation_data$query_cell,]$annotation_prob <- annotation_data$annotation_prob
#  integrated$annotation_level_1 <- factor(tonsil_multi$annotation_level_1)
#}


integrated$annotation_prob <- NA
integrated$annotation_prob[annotation_data$query_cells] <- annotation_data$annotation_prob
integrated$annotation_level_5[annotation_data$query_cells] <- annotation_data$annotation
integrated$UMAP_1_level_5[umap_coords$query_cells] <- umap_coords$UMAP1
integrated$UMAP_2_level_5[umap_coords$query_cells] <- umap_coords$UMAP2
annotation_cite_df <- left_join(
  annotation_data,
  umap_coords,
  by = "query_cells"
)
# Save
saveRDS(integrated, path_to_save_integrated)
saveRDS(annotation_cite_df, path_to_save_cite_seq_df)
```

# Final UMAP

```{r}
color <- c("black", "gray", "red", "yellow", "violet", "green4",
                   "blue", "mediumorchid2", "coral2", "blueviolet",
                   "indianred4", "deepskyblue1", "dimgray", "deeppink1",
                   "green", "lightgray", "hotpink1", "chocolate", "aquamarine", "aliceblue", "burlywood", "blueviolet")
integrated@meta.data %>%
  ggplot(aes(UMAP_1_level_5, UMAP_2_level_5, color = annotation_level_5)) +
    geom_point(size = 0.75) +
    scale_color_manual(values = color) +
    labs(x = "UMAP1", y = "UMAP2", color = "") +
    theme_classic()

```

# Session Information

```{r}
sessionInfo()
```
