==> CITE-seq/08-sub_clustering/CD4_T/04-citeseq_marker_differential.Rmd <==
---
title: "CITEseq marker Differential"
author: "Sonal Rashmi"
date: "`r format(Sys.time(), '%B %Y')`"
output:
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center',
                      message=FALSE, warning = FALSE)
options(width = 1200)
```


# Pre-processing

## Load packages

```{r libraries}
library(Seurat)
library(Signac)
library(flexclust)
library(tidyverse)
library(plyr)
library(harmony)
library(class)
library(ggplot2)
library(reshape2)
library(openxlsx)
```

## Parameters

```{r parameters}
path_to_save_annotated_citeseq <- "/home/srashmi/Documents/tonsil_atlas_citeseq_vdj_20210505/10-sub_cluster_annotation/CD4_T/objects/CD4_tonsil_cite_seq_annotated.rds"
citeseq_marker <- "/home/srashmi/Documents/tonsil_atlas_citeseq_vdj_20210505/10-sub_cluster_annotation/CD4_T/objects/CD4_tonsil_atlas_citeseq_markers.xlsx"
color <- c("black", "gray", "red", "yellow", "violet", "green4",
                   "blue", "mediumorchid2", "coral2", "blueviolet",
                   "indianred4", "deepskyblue1", "dimgray", "deeppink1",
                   "green", "lightgray", "hotpink1", "chocolate", "aquamarine", "aliceblue", "burlywood", "blueviolet")
```

## Variables

## Load data

```{r}
seurat_citeseq <- readRDS(path_to_save_annotated_citeseq)
```


# Find All CITEseq Markers

```{r}
DefaultAssay(seurat_citeseq) <- "ADT"
Idents(seurat_citeseq) <- seurat_citeseq@meta.data$CD4_annotation_level_5
#seurat_citeseq <- NormalizeData(seurat_citeseq, normalization.method = 'CLR', margin = 2) %>% ScaleData()
markers <- FindAllMarkers(seurat_citeseq, assay = "ADT")
```

## Save All Markers in Excel

```{r}
markers <- markers %>%
  dplyr::arrange(cluster, desc(abs(avg_log2FC))) 
#%>% dplyr::filter(p_val_adj < 0.001 & avg_log2FC > 0.9)
markers_list <- purrr::map(levels(markers$cluster), ~ markers[markers$cluster == .x, ])
names(markers_list) <- gsub(":", "-", levels(markers$cluster))
#detach("package:openxlsx", unload = TRUE)
openxlsx::write.xlsx(
  markers_list,
  citeseq_marker
)
#detach("package:openxlsx", unload = TRUE)
DT::datatable(markers, filter ="top")
```


```{r}
seurat_citeseq.markers <- FindAllMarkers(seurat_citeseq, only.pos = TRUE, min.pct = 0.25, logfc.threshold = 0.25)
top_marker_by_clusters <- seurat_citeseq.markers %>% group_by(cluster) %>% top_n(n = 5, wt = avg_log2FC) 

top2 <- seurat_citeseq.markers %>% group_by(cluster) %>% top_n(n = 2, wt = avg_log2FC)

#DoHeatmap(seurat_citeseq, features = top10$gene, label = F)  + RotatedAxis() + NoLegend()
DotPlot(object = seurat_citeseq, features = rev(x =  unique(top2$gene)), cols = c("blue", "red"), dot.scale = 8) + RotatedAxis()
```



# Session Information
```{r session}
sessionInfo()
```
